Small refactoring && Performance optimization
* CommandTag object replaced with string, this object migrated to `pgconn`
* NoticeResponse is no longer inherits ErrorResponse
* Optimized performance and memory usage of ChunkReader
* Optimized performance for BinaryStream read ops
