+++ src/Messages/NoticeResponse.php
@@ -4,8 +4,10 @@ declare(strict_types=1);
 
 namespace PhpPg\PgProto3\Messages;
 
-class NoticeResponse extends ErrorResponse
+class NoticeResponse extends AbstractErrorResponse
 {
+    use MessageName;
+
     public const TYPE = 'N';
 
     public function getType(): string
